Adds question-receiver and moves to api.action/type .
