+++ src/api.action.js
@@ -1,0 +1,10 @@
+import RECEIVE_DATA from "./api.types";
+
+
+export function receiveDataAction (users, questions) {
+  return {
+    type: RECEIVE_DATA,
+    users,
+    questions,
+  }
+}
