--- src/users.action.js
@@ -1,9 +1,0 @@
-import RECEIVE_DATA from "./users.types";
-
-
-export function receiveDataAction (users) {
-  return {
-    type: RECEIVE_DATA,
-    users,
-  }
-}
